Refactor out `Sexp` parser to module
